Update and rename MarkFactory.java to MarkFactory.java
